[Rename] layout 폴더 구분
